move cp performance all sectors chart to use only highcharts react
