+++ app/javascript/components/charts/default-options.js
@@ -1,0 +1,44 @@
+export default {
+  chart: {},
+  xAxis: {
+    title: {
+      text: null
+    },
+    labels: {
+      style: {
+        fontSize: '12px'
+      }
+    }
+  },
+  yAxis: {
+    title: {
+      text: null
+    },
+    labels: {
+      style: {
+        fontSize: '12px'
+      }
+    }
+  },
+  title: {
+    text: null
+  },
+  credits: {
+    enabled: false
+  },
+  legend: {
+    borderWidth: 0
+  },
+  tooltip: {
+    style: {
+      fontSize: '12px'
+    }
+  },
+  plotOptions: {
+    areaspline: {},
+    area: {},
+    series: {
+      marker: {}
+    }
+  }
+};
